move types to Type folder
