+++ src/types/PredictionTypes.ts
